[FIX] mail: fix activity state is shared accross tabs
Since [1], the activity state, which is supposed to be shared
accross tab through a broadcast channel, isn't anymore.

This PR fixes the responsible typo.

[1]: #161286

X-original-commit: 9281f53f55736d04a550546ff2ad37b490d0da8e

+++ addons/mail/static/src/core/web/activity_model_patch.js
@@ -68,16 +68,19 @@ patch(Activity.prototype, {
             [[this.id]],
             { feedback: this.feedback }
         );
-        this.activityBroadcastChannel?.postMessage({
+        this.store.activityBroadcastChannel?.postMessage({
             type: "RELOAD_CHATTER",
             payload: { id: this.res_id, model: this.res_model },
         });
         return action;
     },
     remove({ broadcast = true } = {}) {
+        if (!this.exists()) {
+            return;
+        }
         this.delete();
         if (broadcast) {
-            this.activityBroadcastChannel?.postMessage({
+            this.store.activityBroadcastChannel?.postMessage({
                 type: "DELETE",
                 payload: { id: this.id },
             });
